Create Aleena Vinoji Valliyam Thadathil
Step 1: Input the number to be checked, int a.
Step 2: Check whether the number is 1, 1 is neither prime nor composite.
Step 3: Pass the value to function isPrime(). The variable is read as x in isPrime.
Step 4:Start a for loop, if a number is composite it will be divisible by a number less than or equal to its half thus the loop starts from i=2 to i=x/2.
Step 5: An if is introduced within the for loop with the argument given x%i==0. If the argument is true, 0 is returned to the main function, because the given number is divisible by a number other than 1 and the number itself. 
Step 6: If the if statement work once then the programme comes out of the for loop and goes back to the main function.
Step 7: If the if statement doesn't work throughout the for loop the programme enters the 2nd if statement, and return 1 to the main function.
Step 8: In the main function, the 3rd if statement checks the returned value, if 1 is returned then the given number is prime, and otherwise it is composite.